backend/requirements.txt
Django>=4.0.4
django-cors-headers>=3.12.0
django-model-utils>=4.2.0
dj-database-url>=0.5.0
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.1.0
drf-yasg>=1.20.0
whitenoise>=6.1.0
psycopg2-binary>=2.9.3